requirements.txt
firebase-admin==6.5.0
google-generativeai==0.5.4
protobuf==4.25.3
schedule